Added --g to CLI for GO is_a mapping

+++ goget/cli.py
@@ -3,7 +3,7 @@ Interface CLI commands.
 """
 
 import click
-from dbconn import build_relationships, update_pub_nodes, graph
+from dbconn import build_relationships, update_pub_nodes, graph, create_is_a_cv_term_rel
 from gffproc import examine, parse_gff, get_locus_tags
 from uniprot import query_uniprot
 
@@ -35,9 +35,12 @@ def cli():
               help='Query UniProt using locus tags.')
 @click.option('--p', '--publications', default=True, is_flag=True, prompt='Query PubMed?',
               help='Query PubMed using pmid.')
-def init(gff_file, delete_all, relationships, uniprot, publications):
+@click.option('--g', '--map_go', default=True, is_flag=True, prompt='Map GO Terms?',
+              help='Using QuickGo to map GO is_a relationships.')
+def init(gff_file, delete_all, relationships, uniprot, publications, map_go):
     """
     Load features from GFF file.
+    :param map_go:
     :param publications:
     :param gff_file:
     :param delete_all:
@@ -45,39 +48,43 @@ def init(gff_file, delete_all, relationships, uniprot, publications):
     :param uniprot:
     :return:
     """
-    if delete_all and relationships and not uniprot and not publications:
+    if delete_all and relationships and not uniprot and not publications and not map_go:
         # Deleting existing data, load features and build relationships
         delete_data()
         parse_gff(gff_file)
         build_relationships()
-    elif not delete_all and not uniprot and not publications and relationships:
+    elif not delete_all and not uniprot and not publications and not map_go and relationships:
         # Build relationships from existing data
         build_relationships()
-    elif delete_all and not relationships and not uniprot and not publications:
+    elif delete_all and not relationships and not uniprot and not publications and not map_go:
         # Deleting existing data, load features
         delete_data()
         parse_gff(gff_file)
-    elif delete_all and relationships and uniprot and not publications:
+    elif delete_all and relationships and uniprot and not publications and not map_go:
         # Deleting existing data, load features, build relationships, fetch data from UniProt and create nodes,
         # and build relationships then update Publication nodes with data from PubMed
         delete_data()
         parse_gff(gff_file)
         build_relationships()
         query_uniprot(get_locus_tags(gff_file, 400))
-    elif delete_all and relationships and uniprot and publications:
+    elif delete_all and relationships and uniprot and publications and map_go:
         # Deleting existing data, load features, build relationships, fetch data from UniProt and create nodes,
-        # and build relationships then update Publication nodes with data from PubMed
+        # build relationships then update Publication nodes with data from PubMed and map GO
         delete_data()
         parse_gff(gff_file)
         build_relationships()
         query_uniprot(get_locus_tags(gff_file, 400))
         update_pub_nodes()
-    elif not delete_all and not relationships and not publications and uniprot:
+        create_is_a_cv_term_rel()
+    elif not delete_all and not relationships and not publications and not map_go and uniprot:
         # Fetch data from UniProt and create nodes
         query_uniprot(get_locus_tags(gff_file, 400))
-    elif not delete_all and not relationships and not uniprot and publications:
+    elif not delete_all and not relationships and not uniprot and not map_go and publications:
         # Update Publication nodes with data from PubMed
         update_pub_nodes()
+    elif not delete_all and not relationships and not uniprot and not publications and map_go:
+        # Create is_a relationship between GO using QuickGo
+        create_is_a_cv_term_rel()
     else:
         click.Abort()
